requirements.txt
kafka-python>=2.0.2
kazoo>=2.8.0